Updated changes in capturing captions
